modularized 90% of the commands/events
