+++ package.json
@@ -14,6 +14,7 @@
     "canvas": "github:Automattic/node-canvas",
     "discord.js": "github:hydrabolt/discord.js",
     "express": "^4.16.3",
+    "fs-extra": "^6.0.1",
     "glob": "^7.1.2",
     "image-downloader": "^3.3.0",
     "moment": "^2.22.1",
